saving work-- doing estimator / module split for graphs
