+++ testing/test_samplers_and_trajectories.py
@@ -6,10 +6,11 @@ import torch
 from gfn.containers import Trajectories, Transitions
 from gfn.containers.replay_buffer import ReplayBuffer
 from gfn.gym import Box, DiscreteEBM, HyperGrid
+from gfn.gym.graph_building import RingGraphBuilding
 from gfn.gym.helpers.box_utils import BoxPBEstimator, BoxPBMLP, BoxPFEstimator, BoxPFMLP
-from gfn.modules import DiscretePolicyEstimator, GFNModule
+from gfn.modules import DiscretePolicyEstimator, GFNModule, GNNEdgePolicyModule
 from gfn.samplers import LocalSearchSampler, Sampler
-from gfn.utils.modules import MLP
+from gfn.utils.modules import MLP, GraphEdgeActionGNN
 from gfn.utils.prob_calculations import get_trajectory_pfs
 from gfn.utils.training import states_actions_tns_to_traj
 
@@ -371,63 +372,33 @@ def test_states_actions_tns_to_traj():
     replay_buffer.add(trajs)
 
 
-# ------ GRAPH TESTS ------
+def test_ring_graph_building():
+    env = RingGraphBuilding(
+        n_nodes=10,
+        directed=False,
+        device=torch.device("cpu"),
+        state_evaluator=lambda s: torch.zeros(s.batch_shape),
+    )
+    estimator = GraphEdgeActionGNN(
+        n_nodes=env.n_nodes,
+        directed=env.is_directed,
+        num_conv_layers=1,
+        embedding_dim=128,
+        is_backward=False,
+    )
+    module = GNNEdgePolicyModule(
+        module=estimator,
+        is_backward=False,
+    )
 
-# TODO: This test fails randomly. it should not rely on a custom GraphActionNet.
-# def test_graph_building():
-#     feature_dim = 8
-#     env = GraphBuilding(
-#         feature_dim=feature_dim, state_evaluator=lambda s: torch.zeros(s.batch_shape)
-#     )
+    module.to(device=env.device)
 
-#     module = GraphActionNet(feature_dim)
-#     pf_estimator = GraphActionPolicyEstimator(module=module)
+    sampler = Sampler(estimator=module)
+    trajectories = sampler.sample_trajectories(
+        env,
+        n=7,
+        save_logprobs=True,
+        save_estimator_outputs=False,
+    )
 
-#     sampler = Sampler(estimator=pf_estimator)
-#     trajectories = sampler.sample_trajectories(
-#         env,
-#         n=7,
-#         save_logprobs=True,
-#         save_estimator_outputs=False,
-#     )
-
-#     assert len(trajectories) == 7
-
-
-# class GraphActionNet(nn.Module):
-#     def __init__(self, feature_dim: int):
-#         super().__init__()
-#         self.feature_dim = feature_dim
-#         self.action_type_conv = GCNConv(feature_dim, 3)
-#         self.features_conv = GCNConv(feature_dim, feature_dim)
-#         self.edge_index_conv = GCNConv(feature_dim, 8)
-
-#     def forward(self, states: GraphStates) -> TensorDict:
-#         node_feature = states.tensor.x.reshape(-1, self.feature_dim)
-
-#         if states.tensor.x.shape[0] == 0:
-#             action_type = torch.zeros((len(states), 3))
-#             action_type[:, GraphActionType.ADD_NODE] = 1
-#             features = torch.zeros((len(states), self.feature_dim))
-#         else:
-#             action_type = self.action_type_conv(node_feature, states.tensor.edge_index)
-#             action_type = action_type.reshape(
-#                 len(states), -1, action_type.shape[-1]
-#             ).mean(dim=1)
-#             features = self.features_conv(node_feature, states.tensor.edge_index)
-#             features = features.reshape(len(states), -1, features.shape[-1]).mean(dim=1)
-
-#         edge_index = self.edge_index_conv(node_feature, states.tensor.edge_index)
-#         edge_index = torch.einsum("nf,mf->nm", edge_index, edge_index)
-#         edge_index = edge_index[None].repeat(len(states), 1, 1)
-
-#         return TensorDict(
-#             {
-#                 "action_type": action_type,
-#                 "features": features,
-#                 "edge_index": edge_index.reshape(
-#                     states.batch_shape + edge_index.shape[1:]
-#                 ),
-#             },
-#             batch_size=states.batch_shape,
-#         )
+    assert len(trajectories) == 7
